split Coordinates files
